Move Component class to its own module
